Ubuntu package is a dead end
dh_virtualenv doens't allow to build on launchpad because it needs
to download the python package at build time but launchpad doesn't
have internet connection.
we may want to do a proper Debian package, but ngram is not packaged
yet and I'm not sure I want to maintain that, plus we are python3.4
only for now. So let's keep the install the python way for now.