Do not index housenumbers
Housenumbers take a lot of place in the index. Though they are not
a proper discriminating token: addok only match documents (streets)
and then try to match a housenumber in the document.

This is a bit radical, but may worth it: for 59+62 French deps data,
we reduce the Redis size from 640 to 240.